Fexes in project structure, removing reading/writing to console from domain classes
