+++ src/main/MW/service/InputService.java
@@ -2,5 +2,5 @@ package MW.service;
 
 public interface InputService {
 
-    public String input();
+    String input();
 }
